[SE-0001] Combine modules to service main
