+++ server/SesNotSvc/ServiceMain.cpp
@@ -1,5 +1,24 @@
 #include <Windows.h>
 #include <tchar.h>
+#include "BroadcastTcpServer.h"
+#include "MessageCreator.h"
+
+#define DEFAULT_LISTEN_PORT			13389
+#define	DEFAULT_INTERESTING_PORT	3389
+
+unsigned short ListenPort = DEFAULT_LISTEN_PORT;
+unsigned short InterestingPort = DEFAULT_INTERESTING_PORT;
+
+#ifdef _DEBUG
+#define LOGMESSAGE( str ) OutputDebugString( str );
+#else
+#define LOGMESSAGE( str )
+#endif
+
+typedef struct _ServerAndMessage{
+	BroadcastTcpServer* server;
+	MessageCreator* message;
+} ServerAndMessage, *PServerAndMessage;
 
 SERVICE_STATUS        g_ServiceStatus = { 0 };
 SERVICE_STATUS_HANDLE g_StatusHandle = NULL;
@@ -14,6 +33,11 @@ DWORD WINAPI ServiceWorkerThread(LPVOID lpParam);
 
 int _tmain(int argc, TCHAR *argv[])
 {
+	if (argc == 3) {
+		ListenPort = _tstoi(argv[1]);
+		InterestingPort = _tstoi(argv[2]);
+	}
+
 	SERVICE_TABLE_ENTRY ServiceTable[] =
 	{
 		{ SERVICE_NAME, (LPSERVICE_MAIN_FUNCTION)ServiceMain },
@@ -22,27 +46,33 @@ int _tmain(int argc, TCHAR *argv[])
 
 	if (StartServiceCtrlDispatcher(ServiceTable) == FALSE)
 	{
-		OutputDebugString(_T("SCNS: Main: StartServiceCtrlDispatcher returned error"));
+		LOGMESSAGE(_T("SCNS: Main: StartServiceCtrlDispatcher returned error"));
 		return GetLastError();
 	}
 
-	OutputDebugString(_T("SCNS: Main: Exit"));
+	LOGMESSAGE(_T("SCNS: Main: Exit"));
 	return 0;
 }
-
 
 VOID WINAPI ServiceMain(DWORD argc, LPTSTR *argv)
 {
 	DWORD Status = E_FAIL;
 
-	OutputDebugString(_T("SCNS: ServiceMain: Entry"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Entry"));
+
+	BroadcastTcpServer server;
+	MessageCreator mc(InterestingPort);
+	ServerAndMessage sam;
+	sam.message = &mc;
+	sam.server = &server;
+	server.listen(ListenPort);
 
 	//g_StatusHandle = RegisterServiceCtrlHandler (SERVICE_NAME, ServiceCtrlHandler);
-	g_StatusHandle = RegisterServiceCtrlHandlerEx(SERVICE_NAME, ServiceCtrlHandlerEx, NULL);
+	g_StatusHandle = RegisterServiceCtrlHandlerEx(SERVICE_NAME, ServiceCtrlHandlerEx, &sam);
 
 	if (g_StatusHandle == NULL)
 	{
-		OutputDebugString(_T("SCNS: ServiceMain: RegisterServiceCtrlHandler returned error"));
+		LOGMESSAGE(_T("SCNS: ServiceMain: RegisterServiceCtrlHandler returned error"));
 		goto EXIT;
 	}
 
@@ -57,19 +87,19 @@ VOID WINAPI ServiceMain(DWORD argc, LPTSTR *argv)
 
 	if (SetServiceStatus(g_StatusHandle, &g_ServiceStatus) == FALSE)
 	{
-		OutputDebugString(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
+		LOGMESSAGE(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
 	}
 
 	/*
 	* Perform tasks neccesary to start the service here
 	*/
-	OutputDebugString(_T("SCNS: ServiceMain: Performing Service Start Operations"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Performing Service Start Operations"));
 
 	// Create stop event to wait on later.
 	g_ServiceStopEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
 	if (g_ServiceStopEvent == NULL)
 	{
-		OutputDebugString(_T("SCNS: ServiceMain: CreateEvent(g_ServiceStopEvent) returned error"));
+		LOGMESSAGE(_T("SCNS: ServiceMain: CreateEvent(g_ServiceStopEvent) returned error"));
 
 		g_ServiceStatus.dwControlsAccepted = 0;
 		g_ServiceStatus.dwCurrentState = SERVICE_STOPPED;
@@ -78,7 +108,7 @@ VOID WINAPI ServiceMain(DWORD argc, LPTSTR *argv)
 
 		if (SetServiceStatus(g_StatusHandle, &g_ServiceStatus) == FALSE)
 		{
-			OutputDebugString(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
+			LOGMESSAGE(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
 		}
 		goto EXIT;
 	}
@@ -91,24 +121,24 @@ VOID WINAPI ServiceMain(DWORD argc, LPTSTR *argv)
 
 	if (SetServiceStatus(g_StatusHandle, &g_ServiceStatus) == FALSE)
 	{
-		OutputDebugString(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
+		LOGMESSAGE(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
 	}
 
 	// Start the thread that will perform the main task of the service
 	HANDLE hThread = CreateThread(NULL, 0, ServiceWorkerThread, NULL, 0, NULL);
 
-	OutputDebugString(_T("SCNS: ServiceMain: Waiting for Worker Thread to complete"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Waiting for Worker Thread to complete"));
 
 	// Wait until our worker thread exits effectively signaling that the service needs to stop
 	WaitForSingleObject(hThread, INFINITE);
 
-	OutputDebugString(_T("SCNS: ServiceMain: Worker Thread Stop Event signaled"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Worker Thread Stop Event signaled"));
 
 
 	/*
 	* Perform any cleanup tasks
 	*/
-	OutputDebugString(_T("SCNS: ServiceMain: Performing Cleanup Operations"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Performing Cleanup Operations"));
 
 	CloseHandle(g_ServiceStopEvent);
 
@@ -119,59 +149,29 @@ VOID WINAPI ServiceMain(DWORD argc, LPTSTR *argv)
 
 	if (SetServiceStatus(g_StatusHandle, &g_ServiceStatus) == FALSE)
 	{
-		OutputDebugString(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
+		LOGMESSAGE(_T("SCNS: ServiceMain: SetServiceStatus returned error"));
 	}
 
 EXIT:
-	OutputDebugString(_T("SCNS: ServiceMain: Exit"));
+	LOGMESSAGE(_T("SCNS: ServiceMain: Exit"));
 
 	return;
 }
 
-void NoticeSessionChanging(DWORD message) {
-	switch (message) {
-	case WTS_CONSOLE_CONNECT:
-		OutputDebugString(_T("WTS_CONSOLE_CONNECT"));
-		break;
-	case WTS_CONSOLE_DISCONNECT:
-		OutputDebugString(_T("WTS_CONSOLE_DISCONNECT"));
-		break;
-	case WTS_REMOTE_CONNECT:
-		OutputDebugString(_T("WTS_REMOTE_CONNECT"));
-		break;
-	case WTS_REMOTE_DISCONNECT:
-		OutputDebugString(_T("WTS_REMOTE_DISCONNECT"));
-		break;
-	case WTS_SESSION_LOGON:
-		OutputDebugString(_T("WTS_SESSION_LOGON"));
-		break;
-	case WTS_SESSION_LOGOFF:
-		OutputDebugString(_T("WTS_SESSION_LOGOFF"));
-		break;
-	case WTS_SESSION_LOCK:
-		OutputDebugString(_T("WTS_SESSION_LOCK"));
-		break;
-	case WTS_SESSION_UNLOCK:
-		OutputDebugString(_T("WTS_SESSION_UNLOCK"));
-		break;
-	case WTS_SESSION_REMOTE_CONTROL:
-		OutputDebugString(_T("WTS_SESSION_REMOTE_CONTROL"));
-		break;
-	default:
-		OutputDebugString(_T("NoticeSessionChanging etc"));
-		break;
-	}
+void NoticeSessionChanging(DWORD code, LPVOID lpContext) {
+	ServerAndMessage* sam = (ServerAndMessage*)lpContext;
+	sam->server->broadcast(sam->message->makeMessage(code).c_str());
 }
 
 VOID WINAPI ServiceCtrlHandler(DWORD CtrlCode)
 {
-	OutputDebugString(_T("SCNS: ServiceCtrlHandler: Entry"));
+	LOGMESSAGE(_T("SCNS: ServiceCtrlHandler: Entry"));
 
 	switch (CtrlCode)
 	{
 	case SERVICE_CONTROL_STOP:
 
-		OutputDebugString(_T("SCNS: ServiceCtrlHandler: SERVICE_CONTROL_STOP Request"));
+		LOGMESSAGE(_T("SCNS: ServiceCtrlHandler: SERVICE_CONTROL_STOP Request"));
 
 		if (g_ServiceStatus.dwCurrentState != SERVICE_RUNNING)
 			break;
@@ -187,7 +187,7 @@ VOID WINAPI ServiceCtrlHandler(DWORD CtrlCode)
 
 		if (SetServiceStatus(g_StatusHandle, &g_ServiceStatus) == FALSE)
 		{
-			OutputDebugString(_T("SCNS: ServiceCtrlHandler: SetServiceStatus returned error"));
+			LOGMESSAGE(_T("SCNS: ServiceCtrlHandler: SetServiceStatus returned error"));
 		}
 
 		// This will signal the worker thread to start shutting down
@@ -199,13 +199,14 @@ VOID WINAPI ServiceCtrlHandler(DWORD CtrlCode)
 		break;
 	}
 
-	OutputDebugString(_T("SCNS: ServiceCtrlHandler: Exit"));
+	LOGMESSAGE(_T("SCNS: ServiceCtrlHandler: Exit"));
 }
 
 DWORD WINAPI ServiceCtrlHandlerEx(DWORD dwControl, DWORD dwEventType, LPVOID lpEventData, LPVOID lpContext) {
 
 	if (dwControl == SERVICE_CONTROL_SESSIONCHANGE) {
-		NoticeSessionChanging(dwEventType);
+
+		NoticeSessionChanging(dwEventType, lpContext);
 	}
 	else {
 		ServiceCtrlHandler(dwControl);
@@ -216,7 +217,7 @@ DWORD WINAPI ServiceCtrlHandlerEx(DWORD dwControl, DWORD dwEventType, LPVOID lpE
 
 DWORD WINAPI ServiceWorkerThread(LPVOID lpParam)
 {
-	OutputDebugString(_T("SCNS: ServiceWorkerThread: Entry"));
+	LOGMESSAGE(_T("SCNS: ServiceWorkerThread: Entry"));
 
 	//  Periodically check if the service has been requested to stop
 	while (WaitForSingleObject(g_ServiceStopEvent, 0) != WAIT_OBJECT_0)
@@ -229,7 +230,7 @@ DWORD WINAPI ServiceWorkerThread(LPVOID lpParam)
 		Sleep(3000);
 	}
 
-	OutputDebugString(_T("SCNS: ServiceWorkerThread: Exit"));
+	LOGMESSAGE(_T("SCNS: ServiceWorkerThread: Exit"));
 
 	return ERROR_SUCCESS;
 }
